Update README.md
Background

The goal of this homework is to get some experience building a dataset from online sources, cleaning it and performing some basic visualizations. To do so, you will make use of scraping libraries, regex and matplotlib .

Specifically, we will be working with an audio dataset provided by Google Research described here. However, we will assume it is provided in somewhat of a raw format: a CSV file with YouTube IDs, timestamps and tags.

It will be your job to take this CSV, download the associated audio, format them to only include the relevant segment and clean up the data so it can be used for some downstream ML task (for example, training an audio classifier or a generative model to create similar audio samples).

In addition, you will be visualizing various aspects of the dataset to gain a better understanding of the dataset.